feat: allow creation of optional Epic and Story objects in Prompt struct
The `Prompt` struct now contains `create_epic` and `create_story` functions that return `Option<Epic>` and `Option<Story>` respectively. This change allows for the creation of optional Epic and Story objects.

The `create_epic` and `create_story` functions have been modified to return `Option<Epic>` and `Option<Story>` respectively. The user now has the option to cancel and discard the creation of an Epic or Story by entering "x" when prompted for the name.

The `delete_epic` function has been modified to return `Option<String>` instead of `bool`. The user now has the option to cancel and discard the deletion of an Epic by entering "x" when prompted for confirmation.

The `create_epic` and `create_story` functions in the `Prompt` struct have been updated to handle the creation of optional Epic and Story objects. If an Epic or